+++ src/gnu/CMakeLists.txt
@@ -6,10 +6,12 @@ add_library(
   Erlang.cpp
   Geom.cpp
   HypGeom.cpp
+  isaac64.cpp
   LogNorm.cpp
   MLCG.cpp
   NegExp.cpp
   Normal.cpp
+  nrnisaac.cpp
   nrnran123.cpp
   Poisson.cpp
   Random.cpp
